// src/app/features/ToChucNhanSu/KeKhaiTaiSan/DetailKeKhai/type.ts
export interface KeKhaiTaiSanProps {
  _id: string;
  ssoId: string;
  thoiGianNop: null;
  hoTen: string;
  tenViTriChucDanh: string;
  moTaBienDong: null;
  vanBanLienQuan: any[];
  ghiChu: null;
  ghiChuPheDuyet: null;
  tenChucVu: string;
  tenDonVi: string;
  cacLoaiGiayToCoGiaTri: CacLoaiGiayToCoGiaTri[];
  congTrinhXayDungKhac: any[];
  kimLoaiDaQuy: any[];
  nhaO: any[];
  quyenSuDungDat: any[];
  taiKhoanONuocNgoai: any[];
  taiSanKhacGanLienVoiDat: any[];
  taiSanKhac: any[];
  taiSanONuocNgoai: string;
  tien: any[];
  tongThuNhapGiuaHaiLanKeKhai: any;
  bienDong: { [key: string]: BienDong };
  dotKeKhaiId: string;
  thongTinNhanSuId: string;
  gui: boolean;
  trangThaiDuyet: string;
  createdAt: Date;
  updatedAt: Date;
  dotKeKhaiTaiSan: DotKeKhaiTaiSAN;
  thongTinNhanSu?: ThongTinNhanSu;
}

export interface BienDong {
  giaTri: number | string;
  giaiTrinh: string;
  soLuong: number | string;
}

export interface CacLoaiGiayToCoGiaTri {
  loai: string;
  tenTaiSan: string;
  soLuong: number;
  giaTri: number;
}

export interface DotKeKhaiTaiSAN {
  _id: string;
  tenDot: string;
  urlVanBanDaDongDau: null;
  thoiGianBatDau: Date;
  daGui: null;
  urlQuyetDinhCongVan: null;
  thoiGianKetThuc: Date;
  ghiChu: null;
  vanBanLienQuan: null;
  danhSachKeKhai: DanhSachKeKhai[];
  trangThaiDuyetDot: string;
  soQuyetDinh: null;
  ngayRaQuyetDinh: null;
  nguoiKy: null;
  chucVuNguoiKy: null;
  fileQuyetDinh: null;
  noiDungYeuCauChinhSua: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface DanhSachKeKhai {
  loai: string;
  viTriChucDanhId: string;
  thongTinNhanSuId: string;
  ssoId: string;
}

export interface ThongTinNhanSu {
  _id: string;
  ssoId: string;
  hoDem: string;
  ten: string;
  ngaySinh: Date;
  maCanBo: string;
  tenGoiKhac: string;
  biDanh: null;
  trangThai: string;
  gioiTinh: string;
  email: string;
  noiSinhThanhPhoMa: string;
  noiSinhThanhPhoTen: string;
  noiSinhQuanMa: string;
  noiSinhQuanTen: string;
  noiSinhXaMa: string;
  noiSinhXaTen: string;
  queQuanThanhPhoMa: string;
  queQuanThanhPhoTen: string;
  queQuanQuanMa: string;
  queQuanQuanTen: string;
  queQuanXaMa: string;
  queQuanXaTen: string;
  hoKhauSoNha: string;
  hoKhauThanhPhoMa: string;
  hoKhauThanhPhoTen: string;
  hoKhauQuanMa: string;
  hoKhauQuanTen: string;
  hoKhauXaMa: string;
  hoKhauXaTen: string;
  noiOSoNha: string;
  noiOThanhPhoMa: string;
  noiOThanhPhoTen: string;
  noiOQuanMa: string;
  noiOQuanTen: string;
  noiOXaMa: string;
  noiOXaTen: string;
  sdtCaNhan: string;
  sdtNhaRieng: null;
  sdtCoQuan: null;
  ghiChu: null;
  fullName: null;
  quocTichId: string;
  danTocId: string;
  tonGiaoId: string;
  ngayVaoNganhYTe: null;
  cccdCMND: string;
  ngayCap: Date;
  noiCap: string;
  tinhTrangHonNhanId: string;
  tenNganHang: string;
  chiNhanh: string;
  soTaiKhoan: string;
  soSoBHXH: string;
  noiCapBHXH: null;
  ngayCapBHXH: null;
  ngayThamGiaBHXH: null;
  ghiChuBHXH: null;
  maSoThue: string;
  ngayCapMaSoThue: null;
  listFileUrl: null;
  chieuCao: number;
  canNang: number;
  nhomMau: string;
  tinhTrangSucKhoe: string;
  laGiangVienCoHuu: null;
  chungChiSuPhamGiangVien: null;
  laChuyenVien: null;
  trinhDoGiaoDucPhoThongId: string;
  trinhDoLyLuanChinhTriId: string;
  trinhDoQuanLyHanhChinhId: string;
  trinhDoTinHocId: string;
  danhHieuPhongTangId: null;
  soTruongCongTac: string;
  ngoaiNguId: string;
  khungNangLucNgoaiNguId: string;
  ngayVaoDangDuBi: Date;
  ngayChinhThuc: Date;
  noiVaoDang: string;
  soTheDang: string;
  chucVuDangId: null;
  ngayNhapNgu: Date;
  ngayXuatNgu: Date;
  donViQuanDoi: string;
  chucVuQuanDoi: string;
  ngayVaoDoan: Date;
  noiVaoDoan: string;
  soTheDoan: null;
  chucVuDoanId: null;
  ngayThamGiaCongDoan: null;
  noiThamGiaCongDoan: null;
  soTheCongDoan: null;
  hinhThucTuyenDungId: string;
  dotTuyenDungId: string;
  ngayTuyenDung: Date;
  soBaoDanh: string;
  ngayBatDauLamViec: null;
  ngayVaoNganh: Date;
  emailCanBo: string;
  ngachTuyenDungId: null;
  donViViTriTuyenDungId: string;
  donViTuyenDungId: string;
  anhDaiDien: null;
  loaiHinhLamViec: null;
  trinhDoQuanLyNhaNuocId: string;
  urlAnhDaiDien: null;
  donViQuanLyId: null;
  donViChinhId: string;
  loaiHoSo: string;
  loaiHoSoKhac: null;
  laGiangVienNuocNgoai: null;
  coQuanChuQuan: null;
  soGiayPhepLaoDong: null;
  lichSuBanThanKhaiRo: string;
  lichSuBanThanThamGia: string;
  lichSuBanThanCoThanNhan: string;
  loaiCanBoGiangVien: string;
  chucDanhNgheNghiep: null;
  hang: null;
  maDanToc: string;
  namDuocPhong: null;
  ngayHuong: null;
  tenDanToc: string;
  maQuocTich: string;
  tenQuocTich: string;
  chucVuCongTac: null;
  donViDangCongTac: null;
  idDotCapNhat: null;
  maTonGiao: string;
  tenTonGiao: string;
  soHieuVienChuc: string;
  hocHam: null;
  hocVi: null;
  linhVucDaoTaoId: null;
  linhVucDaoTaoMa: null;
  soHochieu: null;
  toChucCuDen: null;
  thoiGianBatDauCuDen: null;
  thoiGianKetThucCuden: null;
  linhVuc: null;
  chuyenNganh: null;
  linhVucDaoTaoTen: null;
  tenNghiaVuDanQuanTuVe: null;
  ngayKetNapDanQuanTuVe: Date;
  ngayHoanThanhNghiaVu: Date;
  chucDanhDanQuanTuVe: string;
  soQuyetDinhHoanThanh: string;
  nguoiKyQuyetDinh: string;
  maHeThong: number;
  maDonViChinh: string;
  chucVuChinhId: string;
  maChucVuChinh: string;
  chatLuongNhanSu: string;
  namKyHopDong: null;
  maHoSo: null;
  conViTriChucDanhCoHieuLuc: boolean;
  donViViTriId: string;
  maNoiBo: string;
  khac: null;
  tenDonViTuyenDung: string;
  tenDonViViTriTuyenDung: string;
  viTriUngTuyen: null;
  banGoc: boolean;
  chucDanhGiangVienId: null;
  trangThaiChinhSua: string;
  trangThaiHoSoUngTuyen: null;
  ngachLuongId: null;
  hoTen: string;
  thoiGianGuiDuyet: null;
  ngayBatDauLamViecTaiTruong: Date;
  yeuCauChinhSua: null;
  urlFileMinhChung: null;
  taiLieuUngTuyen: null;
  sinhMaHeThong: boolean;
  createdAt: Date;
  updatedAt: Date;
  danhSachQuanHeGiaDinhVeBanThan: DanhSachQuanHeGiaDinhVeBanThan[];
}

export interface DanhSachQuanHeGiaDinhVeBanThan {
  _id: string;
  thongTinNhanSuId: string;
  idGoc: null;
  hoVaTen: string;
  soDienThoai: null;
  namSinh: number;
  moiQuanHe: string;
  noiDung: null;
  nguoiPhuThuoc: boolean;
  ngheNghiep: string;
  noiCongTac: string;
  thangSinh: number;
  ngaySinh: number;
  cccdCMND: null;
  ngayCap: null;
  noiCap: null;
  noiThuongTru: string;
  hoKhauSoNha: null;
  hoKhauThanhPhoMa: string;
  hoKhauThanhPhoTen: string;
  hoKhauQuanMa: string;
  hoKhauQuanTen: string;
  hoKhauXaMa: string;
  hoKhauXaTen: string;
  createdAt: Date;
  updatedAt: Date;
}